requirements.txt
flask==3.1.2
beautifulsoup4==4.12.3
pywinrm==0.4.3
python-dotenv==1.2.1
requests==2.32.3
